+++ tests/testthat/test-kernelshap.R
@@ -1,5 +1,5 @@
 # Model with non-linearities and interactions
-fit <- stats::lm(
+fit <- lm(
   Sepal.Length ~ poly(Petal.Width, degree = 2L) * Species + Petal.Length, data = iris
 )
 x <- c("Petal.Width", "Species", "Petal.Length")
@@ -107,9 +107,9 @@ test_that("feature_names must be in colnames(X) and colnames(bg_X)", {
   expect_error(kernelshap(fit, cbind(iris, a = 1), bg_X = iris, feature_names = "a"))
 })
 
-fit <- stats::lm(Sepal.Length ~ stats::poly(Petal.Width, degree = 2L), data = iris)
+fit <- lm(Sepal.Length ~ poly(Petal.Width, degree = 2L), data = iris)
 x <- "Petal.Width"
-preds <- unname(stats::predict(fit, iris))
+preds <- unname(predict(fit, iris))
 
 test_that("Special case p = 1 works", {
   s <- kernelshap(fit, iris[1:5, x, drop = FALSE], bg_X = iris, verbose = FALSE)
@@ -126,9 +126,9 @@ test_that("Special case p = 1 is chatty with verbose = TRUE", {
   )
 })
 
-fit <- stats::lm(Sepal.Length ~ ., data = iris[1:4])
+fit <- lm(Sepal.Length ~ ., data = iris[1:4])
 X <- data.matrix(iris[2:4])
-pred_fun <- function(m, X) stats::predict(m, as.data.frame(X))
+pred_fun <- function(m, X) predict(m, as.data.frame(X))
 preds <- unname(pred_fun(fit, X))
 s <- kernelshap(fit, X[1:3, ], pred_fun = pred_fun, bg_X = X, verbose = FALSE)
 
@@ -160,18 +160,18 @@ test_that("Matrix input gives error with inconsistent feature_names", {
 
 
 ## Now with case weights
-fit <- stats::lm(
+fit <- lm(
   Sepal.Length ~ poly(Petal.Width, degree = 2L) * Species, data = iris, 
   weights = Petal.Length
 )
 x <- c("Petal.Width", "Species")
-preds <- unname(stats::predict(fit, iris))
+preds <- unname(predict(fit, iris))
 s <- kernelshap(
   fit, iris[1:5, x], bg_X = iris, bg_w = iris$Petal.Length, verbose = FALSE
 )
 
 test_that("Baseline equals weighted average prediction on background data", {
-  expect_equal(s$baseline, stats::weighted.mean(iris$Sepal.Length, iris$Petal.Length))
+  expect_equal(s$baseline, weighted.mean(iris$Sepal.Length, iris$Petal.Length))
 })
 
 test_that("SHAP + baseline = prediction works with case weights", {
@@ -182,17 +182,17 @@ test_that("Decomposing a single row works with case weights", {
   s <- kernelshap(
     fit, iris[1L, x], bg_X = iris, bg_w = iris$Petal.Length, verbose = FALSE
   )
-  expect_equal(s$baseline, stats::weighted.mean(iris$Sepal.Length, iris$Petal.Length))
+  expect_equal(s$baseline, weighted.mean(iris$Sepal.Length, iris$Petal.Length))
   expect_equal(rowSums(s$S) + s$baseline, preds[1L])
 })
 
-fit <- stats::lm(
-  Sepal.Length ~ stats::poly(Petal.Width, degree = 2L), 
+fit <- lm(
+  Sepal.Length ~ poly(Petal.Width, degree = 2L), 
   data = iris, 
   weights = Petal.Length
 )
 x <- "Petal.Width"
-preds <- unname(stats::predict(fit, iris))
+preds <- unname(predict(fit, iris))
 
 test_that("Special case p = 1 works with case weights", {
   s <- kernelshap(
@@ -207,7 +207,7 @@ test_that("Special case p = 1 works with case weights", {
   expect_equal(rowSums(s$S) + s$baseline, preds[1:5])
 })
 
-fit <- stats::lm(
+fit <- lm(
   Sepal.Length ~ . , data = iris[c(1L, 3L, 4L)], weights = iris$Sepal.Width
 )
 X <- data.matrix(iris[3:4])
@@ -235,7 +235,7 @@ s <- kernelshap(fit, X[1L, ], bg_X = X, verbose = FALSE)
 
 test_that("kernelshap works for large p (hybrid case)", {
   expect_equal(s$baseline, mean(y))
-  expect_equal(rowSums(s$S) + s$baseline, unname(stats::predict(fit, X[1L, ])))
+  expect_equal(rowSums(s$S) + s$baseline, unname(predict(fit, X[1L, ])))
 })