+++ tests/testthat/test-methods.R
@@ -1,4 +1,4 @@
-fit <- stats::lm(Sepal.Length ~ ., data = iris)
+fit <- lm(Sepal.Length ~ ., data = iris)
 s <- kernelshap(fit, iris[1:2, -1L], bg_X = iris[-1L], verbose = FALSE)
 
 test_that("is_kernelshap() works", {